Add files via upload
In this project, I analyzed the sales performance of Alpha Cable Ltd for 2022 and 2023. I cleaned the stock report data with Excel, imported the data to Power BI, wrote various DAX to create new tables and measures and modelled the data to create relationships among them. Then I wrote an amazing story with different charts and graphs with Power BI dashboard.
									</p>